Sanitize feed titles
Attempt to fix a recently reported XSS vulnerability.

We reuse the same sanitization strategy as for _story_ titles.

Fixes #490.
